# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Summary Platform API
  version: 1.0.0
  description: >
    음성 파일 업로드, 요약, 다운로드를 지원하는 API 명세서입니다.  
    클라이언트 요청 오류 시 응답 형식은 `{ "detail": "에러 메시지" }`를 따릅니다.

servers:
  - url: http://localhost:8000
    description: Local development server

paths:

  /workspaces:
    post:
      summary: 워크스페이스 생성
      description: 모든 Subject를 담는 최상위 디렉토리를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 개인 프로젝트
      responses:
        '201':
          description: 성공적으로 생성됨
        '409':
          description: 동일한 이름의 워크스페이스가 이미 존재함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 요청 본문 형식이 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: 워크스페이스 목록 조회
      responses:
        '200':
          description: 성공적으로 조회됨
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'

  /subjects:
    post:
      summary: 새로운 항목(디렉토리) 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                is_korean_only:
                  type: boolean
              required: [name]
      responses:
        '201':
          description: 성공적으로 생성됨
        '409':
          description: 동일한 이름의 항목이 이미 존재함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 요청 본문 형식이 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: 생성된 모든 항목 조회
      responses:
        '200':
          description: 성공적으로 조회됨
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'

  /subjects/{subject_id}:
    delete:
      summary: 특정 항목 삭제
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 성공적으로 삭제됨
        '404':
          description: 항목을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /summary-jobs:
    post:
      summary: 요약 작업 생성 및 파일 업로드
      description: 새로운 요약 작업을 생성하고 음성 파일을 업로드합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "선형대수 강의 요약"
                subject_id:
                  type: integer
                files:
                  type: string
                  format: binary
      responses:
        '201':
          description: 성공적으로 작업 생성 및 업로드 시작
        '400':
          description: 잘못된 subject_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: 파일 크기 제한 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: 허용되지 않은 파일 확장자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: 전체 요약 작업 목록 조회
      parameters:
        - in: query
          name: subject_id
          schema:
            type: integer
      responses:
        '200':
          description: 성공적으로 조회됨
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SummaryJob'

  /summary-jobs/{job_id}:
    get:
      summary: 특정 요약 작업 상세 조회
      description: 진행 상태를 포함하여 작업 상세 정보를 반환합니다.  
        프론트엔드는 5초 간격으로 polling하는 것을 권장합니다.
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공적으로 조회됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryJob'
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: 요약 작업 삭제
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공적으로 삭제됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /summary-jobs/{job_id}/download:
    get:
      summary: 완료된 요약 결과 다운로드
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공적으로 파일 반환
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: 작업이 아직 완료되지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 결과 파일을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Job not found"

    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/Subject'

    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        is_korean_only:
          type: boolean
        created_at:
          type: string
          format: date-time

    SummaryJob:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        subject_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
        files:
          type: array
          items:
            $ref: '#/components/schemas/SourceMaterial'

    SourceMaterial:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        status:
          type: string
          enum: [UPLOADED, TRANSCRIBING, SUMMARIZING, COMPLETED, FAILED]
        created_at:
          type: string
          format: date-time

tags:
  - name: Workspace
    description: 워크스페이스 관련 엔드포인트
  - name: Subject
    description: 항목(디렉토리) 관련 엔드포인트
  - name: SummaryJob
    description: 요약 작업 관리 엔드포인트